+++ packages/adapter-netlify/index.js
@@ -9,7 +9,7 @@ export default function () {
 	const adapter = {
 		name: '@sveltejs/adapter-netlify',
 
-		async adapt(utils) {
+		async adapt({ utils }) {
 			const { publish, functions } = validate_config().build;
 
 			utils.rimraf(publish);
@@ -39,6 +39,7 @@ export default function () {
 			utils.copy_client_files(publish);
 
 			utils.log.minor('Writing redirects...');
+			utils.copy('_redirects', `${publish}/_redirects`);
 			appendFileSync(`${publish}/_redirects`, '\n\n/* /.netlify/functions/render 200');
 		}
 	};