+++ packages/adapter-static/index.js
@@ -5,12 +5,12 @@
  *   fallback?: string;
  * }} [opts]
  */
-export default function ({ pages = 'build', assets = pages, fallback = null } = {}) {
+export default function ({ pages = 'build', assets = pages, fallback } = {}) {
 	/** @type {import('@sveltejs/kit').Adapter} */
 	const adapter = {
 		name: '@sveltejs/adapter-static',
 
-		async adapt(utils) {
+		async adapt({ utils }) {
 			utils.copy_static_files(assets);
 			utils.copy_client_files(assets);