+++ packages/kit/src/core/load_config/options.js
@@ -80,6 +80,27 @@ const options = {
 
 			hydrate: expect_boolean(true),
 
+			package: {
+				type: 'branch',
+				children: {
+					dir: expect_string('package'),
+					exports: {
+						type: 'branch',
+						children: {
+							include: expect_array_of_strings(['**']),
+							exclude: expect_array_of_strings(['_*', '**/_*'])
+						}
+					},
+					files: {
+						type: 'branch',
+						children: {
+							include: expect_array_of_strings(['**']),
+							exclude: expect_array_of_strings([])
+						}
+					}
+				}
+			},
+
 			paths: {
 				type: 'branch',
 				children: {
@@ -165,6 +186,23 @@ function expect_string(string, allow_empty = true) {
 			assert_is_string(option, keypath);
 			if (!allow_empty && option === '') {
 				throw new Error(`${keypath} cannot be empty`);
+			}
+			return option;
+		}
+	};
+}
+
+/**
+ * @param {string[]} array
+ * @returns {ConfigDefinition}
+ */
+function expect_array_of_strings(array) {
+	return {
+		type: 'leaf',
+		default: array,
+		validate: (option, keypath) => {
+			if (!Array.isArray(option) || !option.every((glob) => typeof glob === 'string')) {
+				throw new Error(`${keypath} must be an array of strings`);
 			}
 			return option;
 		}