+++ packages/kit/types/endpoint.d.ts
@@ -1,16 +1,5 @@
-import { Headers, ParameterizedBody } from './helper';
-
-export type ServerRequest<Locals = Record<string, any>, Body = unknown> = {
-	method: string;
-	host: string;
-	headers: Headers;
-	path: string;
-	params: Record<string, string>;
-	query: URLSearchParams;
-	rawBody: string | Uint8Array;
-	body: ParameterizedBody<Body>;
-	locals: Locals;
-};
+import { ServerRequest } from './hooks';
+import { Headers, MaybePromise } from './helper';
 
 type JSONValue =
 	| string
@@ -24,9 +13,9 @@ type JSONValue =
 export type EndpointOutput = {
 	status?: number;
 	headers?: Partial<Headers>;
-	body?: string | Uint8Array | JSONValue;
+	body?: JSONValue | Uint8Array;
 };
 
 export type RequestHandler<Locals = Record<string, any>, Body = unknown> = (
 	request: ServerRequest<Locals, Body>
-) => void | EndpointOutput | Promise<EndpointOutput>;
+) => MaybePromise<void | EndpointOutput>;