+++ packages/kit/types/hooks.d.ts
@@ -1,27 +1,33 @@
-import { BaseBody, Headers } from './helper';
-import { ServerRequest } from './endpoint';
+import { Headers, Location, MaybePromise, ParameterizedBody } from './helper';
 
-export type Incoming = {
+export type StrictBody = string | Uint8Array | null;
+
+export type Incoming = Omit<Location, 'params'> & {
 	method: string;
-	host: string;
 	headers: Headers;
-	path: string;
-	query: URLSearchParams;
-	rawBody: string | Uint8Array;
-	body?: BaseBody;
+	rawBody: StrictBody;
+	body?: ParameterizedBody;
+};
+
+export type ServerRequest<Locals = Record<string, any>, Body = unknown> = Location & {
+	method: string;
+	headers: Headers;
+	rawBody: StrictBody;
+	body: ParameterizedBody<Body>;
+	locals: Locals;
 };
 
 export type ServerResponse = {
 	status: number;
 	headers: Headers;
-	body?: string | Uint8Array;
+	body?: StrictBody;
 };
 
 export type GetSession<Locals = Record<string, any>, Session = any> = {
-	(request: ServerRequest<Locals>): Session | Promise<Session>;
+	(request: ServerRequest<Locals>): MaybePromise<Session>;
 };
 
 export type Handle<Locals = Record<string, any>> = (input: {
 	request: ServerRequest<Locals>;
-	render: (request: ServerRequest<Locals>) => ServerResponse | Promise<ServerResponse>;
-}) => ServerResponse | Promise<ServerResponse>;
+	resolve: (request: ServerRequest<Locals>) => MaybePromise<ServerResponse>;
+}) => MaybePromise<ServerResponse>;