+++ packages/adapter-netlify/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sveltejs/adapter-netlify
 
+## 1.0.0-next.17
+
+### Patch Changes
+
+- c51ab7d: Upgrade esbuild to ^0.12.5
+
+## 1.0.0-next.16
+
+### Patch Changes
+
+- edc307d: Remove peerDependencies due to pnpm bug
+- 2636e68: Attempt to fix peerDependencies specification
+- 3b988a4: Allow `_redirects` to be placed in root directory
+
+## 1.0.0-next.15
+
+### Patch Changes
+
+- 028abd9: Pass validated svelte config to adapter adapt function
+- Updated dependencies [6372690]
+- Updated dependencies [c3d36a3]
+- Updated dependencies [bf77940]
+- Updated dependencies [2172469]
+- Updated dependencies [028abd9]
+  - @sveltejs/kit@1.0.0-next.110
+
 ## 1.0.0-next.14
 
 ### Patch Changes